refactor(component): 공통 text 컴포넌트 재설정
 - as prop으로 보다 더 유연하게 태그 설정
 - variant마다 각각 다른 컴포넌트를 반환하지 않고, 그 안의 스타일에서 분기처리 되도록 재설정
 - (minor) 파일 명 변경 (index.styled.ts -> index.styles.ts)
