+++ components/common/Text/index.tsx
@@ -1,12 +1,6 @@
 import theme from '@/styles/theme';
-import React from 'react';
-import {
-  Body,
-  Caption,
-  Heading,
-  PageHeading,
-  SubHeading,
-} from './index.styled';
+import { ElementType, forwardRef, Ref } from 'react';
+import * as Styled from './index.styles';
 
 interface Props
   extends React.HTMLAttributes<
@@ -16,48 +10,33 @@ interface Props
   fontWeight?: 400 | 700;
   color?: keyof typeof theme.colors;
   variant?: 'pageHeading' | 'heading' | 'subHeading' | 'body' | 'caption';
+  as?: ElementType;
 }
 
-const Text = React.forwardRef(
-  ({
-    children,
-    variant = 'body',
-    color = 'gray_90',
-    fontWeight,
-    ...rest
-  }: Props) => {
-    switch (variant) {
-      case 'pageHeading':
-        return (
-          <PageHeading color={color} fontWeight={fontWeight} {...rest}>
-            {children}
-          </PageHeading>
-        );
-      case 'subHeading':
-        return (
-          <SubHeading color={color} fontWeight={fontWeight} {...rest}>
-            {children}
-          </SubHeading>
-        );
-      case 'heading':
-        return (
-          <Heading color={color} fontWeight={fontWeight} {...rest}>
-            {children}
-          </Heading>
-        );
-      case 'caption':
-        return (
-          <Caption color={color} fontWeight={fontWeight} {...rest}>
-            {children}
-          </Caption>
-        );
-      default:
-        return (
-          <Body color={color} fontWeight={fontWeight} {...rest}>
-            {children}
-          </Body>
-        );
-    }
+const Text = forwardRef(
+  (
+    {
+      as,
+      children,
+      variant = 'body',
+      color = 'gray_90',
+      fontWeight,
+      ...rest
+    }: Props,
+    forwardedRef: Ref<HTMLElement>,
+  ) => {
+    return (
+      <Styled.Text
+        as={as}
+        variant={variant}
+        color={color}
+        fontWeight={fontWeight}
+        ref={forwardedRef}
+        {...rest}
+      >
+        {children}
+      </Styled.Text>
+    );
   },
 );
